requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
PyYAML==6.0.3
requests==2.32.5
